Use typed UAVs for Winograd transform shaders
 - Significantly better performance with 4-component vector types.
 - For 320x24 network, dx-fp16 backend is only ~15% slower than cudnn-fp16 on RTX 2060, and fp32 backend is about the same speed.
 - still need to plumb matrix multiply shaders to make it work on AMD.

+++ src/neural/dx/shaders/WinogradTransformSE.hlsl
@@ -99,7 +99,7 @@ void OutputTransformSE
     S = 0;
     float B = 0;
     for (int i = 0; i < se_K; i++) {
-      half val = sharedData[i];
+      float val = sharedData[i];
       S += val * readw2(i, k);
       B += val * readw2(i, k + C);
     }
